+++ e2e/mocks/api.controller.ts
@@ -1,11 +1,18 @@
 import { Controller } from '@nestjs/common';
 import { RMQService } from '../../lib';
-import { DivideContracts, MultiplyContracts, NotificationContracts, SumContracts, TimeOutContracts } from '../contracts/mock.contracts';
+import {
+	AppIdContracts,
+	DivideContracts,
+	ManualAckContracts,
+	MultiplyContracts,
+	NotificationContracts,
+	SumContracts,
+	TimeOutContracts,
+} from '../contracts/mock.contracts';
 
 @Controller()
 export class ApiController {
-	constructor(private readonly rmq: RMQService) {
-	}
+	constructor(private readonly rmq: RMQService) {}
 
 	async sumSuccess(arrayToSum: number[]): Promise<SumContracts.Response> {
 		return this.rmq.send<SumContracts.Request, SumContracts.Response>(SumContracts.topic, { arrayToSum });
@@ -24,24 +31,27 @@ export class ApiController {
 	}
 
 	async multiply(arrayToMultiply: number[]): Promise<MultiplyContracts.Response> {
-		return this.rmq.send<MultiplyContracts.Request, MultiplyContracts.Response>(
-			MultiplyContracts.topic,
-			{ arrayToMultiply },
-		);
+		return this.rmq.send<MultiplyContracts.Request, MultiplyContracts.Response>(MultiplyContracts.topic, {
+			arrayToMultiply,
+		});
 	}
 
-    async timeOutMessage(num: number): Promise<number> {
-       return this.rmq.send<number, number>(
-            TimeOutContracts.topic,
-            num,
-            { timeout: 4000 },
-       )
-    }
+	async timeOutMessage(num: number): Promise<number> {
+		return this.rmq.send<number, number>(TimeOutContracts.topic, num, { timeout: 4000 });
+	}
 
 	async divide(first: number, second: number): Promise<DivideContracts.Response> {
-		return this.rmq.send<DivideContracts.Request, DivideContracts.Response>(
-			DivideContracts.topic,
-			{ first, second },
-		);
+		return this.rmq.send<DivideContracts.Request, DivideContracts.Response>(DivideContracts.topic, {
+			first,
+			second,
+		});
+	}
+
+	async appId(): Promise<AppIdContracts.Response> {
+		return this.rmq.send<null, AppIdContracts.Response>(AppIdContracts.topic, null);
+	}
+
+	async manualAck(): Promise<ManualAckContracts.Response> {
+		return this.rmq.send<null, ManualAckContracts.Response>(ManualAckContracts.topic, null);
 	}
 }