+++ lib/interfaces/queue-meta.interface.ts
@@ -1,5 +1,10 @@
-export interface IQueueMeta {
+export interface IRouteMeta {
 	topic: string;
 	methodName: string;
 	target: any;
+	options?: IRouteOptions;
+}
+
+export interface IRouteOptions {
+	manualAck: boolean;
 }